Add first version
Commit, merge, rebase, and reset subcommands all undo their respective
git commands.
Calling ohshit without any subcommands will try to guess from the reflog
what was done last and try to fix it.
In most cases, since it is impossible to guess what the "intended" goal
was with a command, it is simplest to just "undo" the previous command.